Perf improvements
f64 -> f32 (requires manually implementing simplex noise)

+++ src/palette.rs
@@ -4,14 +4,14 @@ pub struct Rgb {
     pub b: u8,
 }
 
-pub fn hsl_to_rgb(h: f64, s: f64, l: f64) -> Rgb {
+pub fn hsl_to_rgb(h: f32, s: f32, l: f32) -> Rgb {
     let a = s * l.min(1.0 - l);
-    let f = |n: f64| {
+    let f = |n: f32| {
         let k0 = n + h / 30.0;
         let k = if k0 >= 12.0 { k0 - 12.0 } else { k0 };
         l - a * (k - 3.0).min(9.0 - k).min(1.0).max(-1.0)
     };
-    let to_u8 = |v: f64| (v * 255.0).round().clamp(0.0, 255.0) as u8;
+    let to_u8 = |v: f32| (v * 255.0).round().clamp(0.0, 255.0) as u8;
 
     Rgb {
         r: to_u8(f(0.0)),
